serial: 8250_dw: Use a DT prop to stop changing the baud clock
Commit 6a171b29937984a5e0bf29d6577b055998f03edb ("serial: 8250_dw: Allow
hardware flow control to be used") introduced a problem for devices that
cannot generate clock rates that are 16 * baud rate, This was fixed by commit
d8954f3c203824aa84db50abd340990d556ba8e6 ("serial: 8250_dw: Improve clock
rate setting"), however it does not fix it for all use cases.

On SoCs, it is common to have a baud clock that is very high that works for
all clock rates that the device is intended to be used with. This clock is
used for multiple UARTs and so changing the clock rate may cause a different
UART to fail. However, knowledge about shared use of the clock is not known
during probe.

Therefore, this change introduces a DT property (dont-change-baud-clock) that
tells the driver to not change the baud clock rate at all.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>