dma: dw: Handle case when peripheral is the flow controller
If the peripheral is the flow controller, the transfer must not be split
into multiple blocks. The DMAC can support any length transfers in this
mode of operation.

This was found when performing a write of 4093 or 4095 bytes to a SPI FRAM
device. In this case, the SPI operation adds 4 bytes for the command and
address. Since the transfer length is over the max block size supported by
the DMA Controller (4095), the driver creates two DMA descriptors, and the
write fails. Writing 4094 bytes was ok because the DM driver used 16-bit
transfers, hence the number of transfers is lower than the max supported.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>