net: stmmac: Make the software reset optional
There are rare cases where another core has already enabled the GMAC PTP settings
and we do not want to clobber them. This is because the PTP settings affect
the ptp_timestamp signals from the GMAC that can be used by other IP blocks.

For example, on RZ/N1 the ptp_timestamp from GMAC1 is used by GMAC2, the 5-Port
Switch and the HSR Switch.

Signed-off-by: Phil Edworthy <phil.edworthy@renesas.com>