RZ/N1 5-Port Switch workaround (semaphore/spinlock)
Every access to the GMACs, the 5-port Switch and the RIN Engine must be
wrapped in a system-wide mutex to ensure that only one core, including
the Cortex M3, can access the registers at a time.
This is to allow writes to the 5-Port Switch to include a write to a
dummy register immediately before the real register write.

The system-wide mutex is implemented using either:
- the Semaphore hardware on the RZ/N1 device
- a spinlock under Linux kernel
One option must be selected in kernel configuration.

Signed-off-by: Ralph Siemsen <ralph.siemsen@linaro.org>
Signed-off-by: Gareth Williams <gareth.williams.jx@renesas.com>