serial: 8250_dw: RZ/N1 DMA Improvements
The patch enhances RZ/N1 DMA driver

There are some points to note about DMA burst sizes.
First, DMA must not remove all of the data from the rx FIFO. Otherwise, we do not
get a 'character timeout' interrupt, and so do not know that we should push data up
the serial stack. Therefore, we have the rx threshold for generating an interrupt
set to half the FIFO depth (this is the default for 16550A), and set the DMA burst
size when reading the FIFO to a quarter of the FIFO depth.

Second, when transmitting data using DMA, the burst size must be limited to 1 byte
to handle then case when transmitting just 1 byte. Otherwise the DMA doesn't
complete the burst, and nothing happens.

Signed-off-by: Phil Edworthy's avatarPhil Edworthy <phil.edworthy@renesas.com>
Signed-off-by: Biju Das <biju.das.jz@bp.renesas.com>